sync, beware, 1.5 has keywords now
(Portage version: 2.2.01.19833-prefix/svn/Darwin i386, signed Manifest commit with key C5C74E89)

+++ sys-devel/gcc-config/gcc-config-1.4.1-r1.ebuild
@@ -1,11 +1,11 @@
 # Copyright 1999-2011 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-devel/gcc-config/gcc-config-1.5-r1.ebuild,v 1.3 2011/11/11 12:15:40 aballier Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-devel/gcc-config/gcc-config-1.4.1-r1.ebuild,v 1.1 2011/04/28 23:27:02 halcy0n Exp $
 
 inherit eutils flag-o-matic toolchain-funcs multilib prefix
 
 # Version of .c wrapper to use
-W_VER="1.5.2"
+W_VER="1.5.1"
 
 DESCRIPTION="Utility to change the gcc compiler being used"
 HOMEPAGE="http://www.gentoo.org/"
@@ -52,14 +52,12 @@ pkg_postinst() {
 	# Make sure old versions dont exist #79062
 	rm -f "${EROOT}"/usr/sbin/gcc-config
 
-	# We not longer use the /usr/include/g++-v3 hacks, as
-	# it is not needed ...
-	[[ -L ${EROOT}/usr/include/g++ ]] && rm -f "${EROOT}"/usr/include/g++
-	[[ -L ${EROOT}/usr/include/g++-v3 ]] && rm -f "${EROOT}"/usr/include/g++-v3
-
 	# Do we have a valid multi ver setup ?
-	local x
-	for x in $(gcc-config -C -l 2>/dev/null | awk '$NF == "*" { print $2 }') ; do
-		gcc-config ${x}
-	done
+	if gcc-config --get-current-profile &>/dev/null ; then
+		# We not longer use the /usr/include/g++-v3 hacks, as
+		# it is not needed ...
+		[[ -L ${EROOT}/usr/include/g++ ]] && rm -f "${EROOT}"/usr/include/g++
+		[[ -L ${EROOT}/usr/include/g++-v3 ]] && rm -f "${EROOT}"/usr/include/g++-v3
+		gcc-config $(${EPREFIX}/usr/bin/gcc-config --get-current-profile)
+	fi
 }
